# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array"
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WarpPatchInputBody": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "license": {
            "description": "Ключ WARP+ для зарегистрированного аккаунта",
            "maxLength": 40,
            "type": "string"
          },
          "routes": {
            "description": "Домены (example.com — вместе с поддоменами) и сети (104.16.0.0/13)",
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "WarpView": {
        "additionalProperties": false,
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "endpoint": {
            "type": "string"
          },
          "inbounds": {
            "description": "Подключения ноды, у которых весь трафик идёт через WARP",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "ipv4": {
            "type": "string"
          },
          "ipv6": {
            "type": "string"
          },
          "plus": {
            "description": "Аккаунт WARP+",
            "type": "boolean"
          },
          "routes": {
            "description": "Домены и сети, которые идут через WARP у всех подключений ноды",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "source": {
            "enum": [
              "register",
              "import",
              ""
            ],
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WarpViewStatusStruct",
            "description": "Последняя проверка выхода через WARP с ноды; нет — нода недоступна"
          }
        },
        "required": [
          "configured",
          "enabled",
          "plus",
          "routes",
          "inbounds"
        ],
        "type": "object"
      },
      "WarpViewStatusStruct": {
        "additionalProperties": false,
        "properties": {
          "checked_at": {
            "format": "date-time",
            "type": "string"
          },
          "colo": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "ip": {
            "description": "Адрес, который видят сайты",
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "warp": {
            "description": "on | plus | off — как отвечает Cloudflare",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "checked_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "apiKey": {
        "description": "Ключ API: Authorization: Bearer mk_…",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "description": "Сессия админки + заголовок X-CSRF-Token на изменяющих запросах",
        "in": "cookie",
        "name": "__Host-prototip",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "REST API панели prototip. Все пути — под секретным адресом админки: https://<панель>/<секретный путь>/api/v1/…\n\nСкрипты и интеграции авторизуются ключом API (Настройки → API): заголовок `Authorization: Bearer mk_…`. Ключ «чтение» выполняет только GET и не получает ссылок подписок и секретных адресов; «полный» меняет данные, кроме входа, сессий, самих ключей и операций, где уходят деньги, ключи и адреса клиентов (они помечены «только сессия»).\n\nАдминка в браузере ходит с cookie сессии; изменяющие запросы тогда требуют заголовок `X-CSRF-Token` из `GET /auth/me`.\n\nОшибки — RFC 9457 (application/problem+json): `detail` — код ошибки, `errors[].message` — код по полю.",
    "title": "prototip",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/nodes/{id}/warp": {
      "delete": {
        "operationId": "delete-node-warp",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Удалить WARP ноды",
        "tags": [
          "node"
        ]
      },
      "get": {
        "operationId": "get-node-warp",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarpView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "WARP ноды",
        "tags": [
          "node"
        ]
      },
      "patch": {
        "operationId": "update-node-warp",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarpPatchInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarpView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Включить WARP, списки доменов, ключ WARP+",
        "tags": [
          "node"
        ]
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "session": []
    }
  ]
}